requirements.txt
discord.py
flask>=3.0.0
textblob>=0.17.1
transformers>=4.40.0
python-dotenv>=1.0.0
requests>=2.32.0
torch
openai>=1.40.0